requirement.txt
fastapi[all]
aioredis
cos-python-sdk-v5
PyJWT
python-dotenv
tortoise-orm[aiomysql]
tencentcloud-sdk-python
qcloud-python-sts
wechatpy
qrcode[pil]
python-alipay-sdk
passlib
#sqlmodel